Persistent store in cache provider
