+++ src/controllers/open-data-datasets.js
@@ -2,10 +2,10 @@
  * Gets the datasets from the cache and returns as a response
  */
 module.exports = async function(req, res) {
-  let cache = req.cache
+  let store = req.store
   let datasets
   try {
-    datasets = await cache.getDatasetData()
+    datasets = await store.getDatasetData()
   } catch (e) {
     e.code = 500
     throw e
